Moved cycle dependencies checking to main fun
