+++ homework/shared_ptr/CMakeLists.txt
@@ -40,7 +40,7 @@ add_executable(${PROJECT_NAME} main.cpp)
 target_compile_options(${PROJECT_NAME} PRIVATE ${FLAGS})
 target_link_options(${PROJECT_NAME} PRIVATE ${SANITS})
 
-add_executable(${PROJECT_NAME}-ut shared_ptr-ut.cpp sharedControlBlock-ut.cpp make_shared-ut.cpp weak_ptr-ut.cpp shared_weak_coop-ut.cpp)
+add_executable(${PROJECT_NAME}-ut shared_ptr-ut.cpp sharedControlBlock-ut.cpp make_shared-ut.cpp weak_ptr-ut.cpp)
 target_compile_options(${PROJECT_NAME}-ut PRIVATE ${FLAGS})
 target_link_libraries(${PROJECT_NAME}-ut PUBLIC gtest_main)
 target_link_options(${PROJECT_NAME}-ut PRIVATE ${SANITS})
