--- homework/shared_ptr/shared_weak_coop-ut.cpp
@@ -1,17 +1,0 @@
-#include <gtest/gtest.h>
-
-#include "shared_ptr.hpp"
-#include "weak_ptr.hpp"
-
-struct Node {
-    cs::shared_ptr<Node> child;
-    cs::weak_ptr<Node> parent;
-};
-
-TEST(SharedWeakCoopTest, shouldSolveCycleDependecies) {
-    cs::shared_ptr<Node> root(new Node{});
-    cs::shared_ptr<Node> child(new Node{});
-
-    root->child = child;
-    child->parent = root;
-}
